+++ lib/src/plugins/plugins.dart
@@ -1,2 +1,4 @@
-export 'align_ruler/align_ruler.dart' show ManaAlignRuler;
-export 'demo/demo.dart' show Demo;
+export 'demo/index.dart' show Demo;
+export 'mana_align_ruler/index.dart' show ManaAlignRuler;
+export 'mana_logger/index.dart' show ManaLogger;
+export 'mana_logger/mana_logger_collector.dart' show ManaLoggerCollector;