+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0-beta.2
+
+- Add plugin mana_align_ruler
+- Add plugin mana_logger
+
 ## 1.0.0-beta.1
 
 - init